+++ application/controllers/Stock_History.php
@@ -14,28 +14,43 @@ class Stock_History extends MY_Controller {
     }
 
     /**
-     * Generates the stock history page values and passes them to the view
+     * Generates the default stock history page with
      */
     function index() {
-        $this->load->model('StocksModel');
         $this->load->library('table');
 
         $this->data['pagetitle'] = 'Stock History';
         $this->data['pagebody'] = 'stock_history';
-        $value = $this->input->post('stock_select');
+        $this->data['stocks'] = $this->generateDropdown();
+        $mostRecent = $this->StockHistory->getMostRecentStock();
+        $stockHistory = $this->StockHistory->getStockHistory($mostRecent[0]['code']);
+        $this->data['table'] = $this->generateTable($stockHistory);
 
-        $table = NULL;
-        if(!empty($value)) {
-            $stockHistory = $this->StocksModel->getStockHistory($value);
-            $table = $this->generateTable($stockHistory);
-        } else {
-            $mostRecent = $this->StocksModel->getMostRecentStock();
-            $stockHistory = $this->StocksModel->getStockHistory($mostRecent[0]['code']);
-            $table = $this->generateTable($stockHistory);
-        }
-
-        $this->data['table'] = $table;
         $this->render();
+    }
+
+    /**
+     * Displays a given stock history information into the table
+     * @param $stock The stock to display the history about
+     */
+    function displayStock($stock) {
+        $this->load->library('table');
+        $this->data['pagetitle'] = 'Stock History';
+        $this->data['pagebody'] = 'stock_history';
+        $this->data['stocks'] = $this->generateDropdown();
+        $stockHistory = $this->StockHistory->getStockHistory($stock);
+        $this->data['table'] = $this->generateTable($stockHistory);
+
+        $this->render();
+    }
+
+    /**
+     * Routing function for the specific stock
+     */
+    function getSpecificStock() {
+        $stock = $this->input->get('stock_select');
+        redirect("stock-history/$stock");
+
     }
 
     /**
@@ -44,6 +59,10 @@ class Stock_History extends MY_Controller {
      * @return mixed The generated table
      */
     function generateTable($stock) {
+        //sets the table style
+        $template = array('table_open' => '<table class="highlight">');
+        $this->table->set_template($template);
+
         $this->table->set_heading('Code', 'Date', 'Stock Movement',
             'Quantity', 'Transaction');
         foreach ($stock as $row) {
@@ -51,5 +70,20 @@ class Stock_History extends MY_Controller {
         }
 
         return $this->table->generate();
+    }
+
+    /**
+     * Generates the drop down of stock values
+     * @return string The html dropdown
+     */
+    function generateDropdown() {
+        $options = "";
+        $stocks = $this->StockHistory->getStocks();
+        foreach ($stocks as $row) {
+            echo $row['code'];
+            $options .= "<option value=" . $row['code'] . ">" . $row['name'] .
+                "</option>";
+        }
+        return $options;
     }
 }