+++ application/views/home.php
@@ -1,79 +1,38 @@
+<div class="row">
+    <h2 class="left thin">STOCKS</h2>
+</div>
 
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
+<div class="row">
+{stockList}
+<div class="col s12 m6 l4 left">
+    <div class="card blue lighten-2">
         <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
+            <span class="card-title">
+                <a class="yellow-text" href="/stock-history/{code}">{name}</a>
+            </span>
+            <span class="card-title right">{value}</span>
         </div>
     </div>
 </div>
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
-        <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
-        </div>
-    </div>
+{/stockList}
 </div>
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
-        <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
-        </div>
-    </div>
+
+<div class="row">
+    <h2 class="left thin">PLAYERS</h2>
 </div>
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
-        <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
-        </div>
-    </div>
-</div>
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
-        <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
+
+<div class="row">
+{playerList}
+    <div class="col s12 m6 l4 left">
+        <div class="card green lighten-2">
+            <div class="card-content white-text">
+                <span class="card-title">
+                    <a href="/player-portfolio/{Player}">{Player}</a>
+                </span>
+                <span class="card-title right">{equity}</span>
+                <span class="card-title right">{Cash}</span>
+            </div>
         </div>
     </div>
-</div>
-<div class="col s12 m6 l4">
-    <div class="card blue-grey darken-1">
-        <div class="card-content white-text">
-            <span class="card-title">Card Title</span>
-            <p>I am a very simple card. I am good at containing small bits of information.
-            I am convenient because I require little markup to use effectively.</p>
-        </div>
-        <div class="card-action">
-            <a href="#">This is a link</a>
-            <a href="#">This is a link</a>
-        </div>
-    </div>
+{/playerList}
 </div>